pyproject.toml
[tool.poetry]
name = "globus-mas"
version = "0.1.0"
description = ""
authors = ["Uriel <uriel@globus.org>"]

[tool.poetry.dependencies]
python = "^3.8"
globus_sdk = ">1.9.1"
pyyaml = "^5.3.1"
typer = {extras = ["all"], version = "^0.3.2"}
pygments = "^2.7.0"
pydantic = "^1.8.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
isort = "^5.8.0"

[tool.poetry.scripts]
globus-mas = "globus_mas.cli.main:app"
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.isort]
profile = "black"
multi_line_output = 3
